backend/requirements.txt
pandas
openpyxl
django
djangorestframework
django-cors-headers
requests